requirements.txt
sqlalchemy==2.0.45
python-dotenv==1.2.1
requests==2.32.5
redis==7.1.0
geoip2==5.2.0
rosu-pp-py==3.1.0
sendgrid==6.12.5
boto3<2
psycopg2-binary==2.9.11
fastapi==0.124.4
uvicorn==0.38.0
bcrypt==5.0.0
python-multipart==0.0.21
pillow==12.0.0
email-validator==2.3.0
pyjwt==2.10.1
regex==2025.11.3
websockets==15.0.1
orjson==3.11.5
pydantic-settings==2.12.0